requirements.txt
asgiref==3.6.0
Django==4.1.7
django-phonenumber-field==7.0.2
django-phonenumbers==1.0.1
phonenumbers==8.13.8
Pillow==9.4.0
pytz==2023.3
sqlparse==0.4.3
